image_filter_app/requirements.txt
Flask
Werkzeug
Pillow

